Update patch set 2
Patch Set 2:

To the best of my knowledge, yes. It was originally written for the G variant, but it it also resolves wifi not turning on in the P variant as well.

Admittedly, I'm not 100% clear how this is fixing the wifi. My understanding of C and kernel programming is limited. The change creates a static int which appears to be creating a pointer that stores the androidboot.revision value, then uses the printk function to print this information to the syslog.

Since the new code is only printing a new line to syslog, I can't think of any situation where this would negatively interfere with non-G or P variants.

Patch-set: 2